requirements.txt
gradio>=4.0.0
torch>=1.9.0
transformers>=4.20.0
pillow>=9.0.0
numpy>=1.21.0
huggingface-hub>=0.16.0
